backend/pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["PieterT2000 <piet@mapmortar.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
anyio = "4.4.0"
apispec = "6.6.1"
attrs = "23.2.0"
bech32 = "1.2.0"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
cosmpy = "0.9.2"
distro = "1.9.0"
ecdsa = "0.18.0"
frozenlist = "1.4.1"
googleapis-common-protos = "1.63.1"
grpcio = "1.64.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
lxml = "5.2.2"
msgpack = "1.0.8"
multidict = "6.0.5"
openai = "1.33.0"
packaging = "24.1"
pillow = "10.3.0"
protobuf = "4.25.3"
pycryptodome = "3.20.0"
pydantic = "1.10.16"
python-dateutil = "2.9.0.post0"
python-docx = "1.1.2"
python-dotenv = "1.0.1"
referencing = "0.35.1"
requests = "2.32.3"
rpds-py = "0.18.1"
setuptools = "70.0.0"
six = "1.16.0"
sniffio = "1.3.1"
tqdm = "4.66.4"
typing-extensions = "4.12.2"
urllib3 = "2.2.1"
uvicorn = "0.20.0"
websockets = "10.4"
wheel = "0.43.0"
yarl = "1.9.4"
flask = "3.0.3"
Flask-Cors = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
